pyproject.toml
[project]
name = "excel-agent"
version = "0.1.0"
description = "Excel 智能问数 Agent - 基于 LangGraph 的 Excel 数据分析助手"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.9",
    "chromadb>=0.4.0",
    "xlrd>=2.0.1",
    "msoffcrypto-tool>=5.0.0",
]

[project.scripts]
excel-agent = "excel_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/excel_agent"]
